rename `willAnswer` page
this is to reflect the name of the page rather than the section
