+++ server/form-pages/apply/about-the-person/equality-diversity-monitoring/index.ts
@@ -1,12 +1,12 @@
 /* istanbul ignore file */
 
 import { Task } from '../../../utils/decorators'
-import EqualityAndDiversity from './equalityAndDiversity'
+import WillAnswer from './willAnswer'
 import Disability from './disability'
 
 @Task({
   name: 'Complete equality and diversity monitoring',
   slug: 'equality-and-diversity-monitoring',
-  pages: [EqualityAndDiversity, Disability],
+  pages: [WillAnswer, Disability],
 })
 export default class EqualityAndDiversityMonitoring {}
